gnupg: fix keyserver configuration
GnuPG only honors the last given keyserver[0]. So specifying multiple
keyservers does nothing. Furthermore, hkps requires the certificate file
to be specified[1], which may or may be installed. IPv6 may also cause
problems, and gpg does not retry if a connection fails.

For these reasons, we use the IPv4 only pool of sks.

[0] https://lists.gnupg.org/pipermail/gnupg-users/2003-May/018147.html
[1] https://sks-keyservers.net/overview-of-pools.php