weechat: disable password salting
weechat unfortunately changes the salt each time it writes to sec.conf,
even if the password itself hasn't changed. From the documentation of
sec.crypt.salt:

description: use salt when generating key used in encryption
(recommended for maximum security); when enabled, the content of crypted
data in file sec.conf will be different on each write of the file; if
you put the file sec.conf in a version control system, then you can turn
off this option to have always same content in file

Since we're using a long and random password anyway, that isn't being
reused, this is safe to disable.