vim: replace ctrlp with fzf
CtrlP is nice when there are not a lot of files in the directory, but
becomes really slow on larger codebases (i.e. the linux kernel). While
it's possible to work around this to some extend, there are much better
alternatives now.

The most general (and fastest) of which is fzf. Combined with fd, a
faster `find` it handles even absurd amounts of files practically
instantly.

The only real downside of this setup is that it requires both fzf and fd
to be installed.