zsh: limit history in memory to 10000 lines
10000 lines should suffice in general, no need to waste additional
memory. Additionally hist_expire_dups_first has been removed, because
this causes a lot of additional CPU load, probably loading $HISTFILE
over and over again.

If further history is required, it can be loaded with a function like
the following:

function hist {
    local TEMPHISTFILE="$HISTFILE"
    local TEMPSIZE=$SAVEHIST
    fc -pa
    HISTSIZE=$TEMPSIZE
    fc -R "$TEMPHISTFILE"
    fc -l 0
}

The problem with the above is that it will increase the memory footprint
of the zsh process, up to ~50MiB, that doesn't seem to be freed later.

Just grepping $HISTFILE produces a similar result, without this
drawback.